refactor : css 및 js 코드 위치 이동.
